+++ src/problem/SetCover.kt
@@ -1,22 +1,25 @@
 package problem
 
+import data.Subsets
 import data.newName
 import discrete.Goal
 import discrete.Problem
 import discrete.Solution
+import fn.ScoreFn
 import fn.asSubset
 
 fun newSetCover(variant: String, n: Int): Problem? {
 	val name = newName(SetCover, variant, n)
 	return when (variant) {
 		"basic" -> setCover(name)
+		"weighted" -> weightedSetCover(name)
 		else -> null
 	}
 }
 
-fun setCover(name: String): Problem? {
+fun newSetCoverProblem(name: String): Pair<Problem?, Subsets?> {
 	val (p, cfg) = newSubsetsProblem(name)
-	if(p == null || cfg == null) return null
+	if(p == null || cfg == null) return Pair(null, null)
 
 	p.goal = Goal.MINIMIZE
 	p.addUniversalConstraint(fun(solution: Solution): Boolean {
@@ -28,5 +31,19 @@ fun setCover(name: String): Problem? {
 		}
 		return covered.values.all { it }
 	})
+	return Pair(p, cfg)
+}
+
+fun setCover(name: String): Problem? {
+	val (p, _) = newSetCoverProblem(name)
+	return p
+}
+
+fun weightedSetCover(name: String): Problem? {
+	val (p, cfg) = newSetCoverProblem(name)
+	if(p == null || cfg == null) return null
+	if(cfg.weight.size != cfg.names.size) return null
+
+	p.objectiveFn = ScoreFn.sumWeightedSubset(cfg.names, cfg.weight)
 	return p
 }